adds support for custom reporters
