+++ packages/cspell-json-reporter/README.md
@@ -1,0 +1,21 @@
+# `@cspell/cspell-json-reporter`
+
+> CSpell reporter with JSON output
+
+## Usage
+
+Add this to cSpell.json:
+
+```
+reporters: [
+    ['@cspell/cspell-json-reporter', { outFile: 'out.json' }]
+]
+```
+
+## Output file format
+
+TODO
+
+## Settings
+
+TODO
