+++ packages/cspell-json-reporter/src/utils/validateSettings.ts
@@ -1,0 +1,26 @@
+import { CSpellJSONReporterSettings } from '../CSpellJSONReporterSettings';
+
+function assertBooleanOrUndefined(key: string, value: unknown): asserts value is boolean | undefined {
+    if (typeof value !== 'boolean' && value !== undefined) {
+        throw new TypeError(`cspell-json-reporter settings.${key} must be a boolean`);
+    }
+}
+
+/**
+ * Throws an error if passed cspell-json-reporter settings are invalid
+ */
+export function validateSettings(settings: unknown): asserts settings is CSpellJSONReporterSettings {
+    if (!settings || typeof settings !== 'object') {
+        throw new TypeError('cspell-json-reporter settings must be an object');
+    }
+
+    const { outFile, debug, verbose, progress } = settings as CSpellJSONReporterSettings;
+
+    if (typeof outFile !== 'string') {
+        throw new TypeError('cspell-json-reporter settings.outFile must be a string');
+    }
+
+    assertBooleanOrUndefined('verbose', verbose);
+    assertBooleanOrUndefined('debug', debug);
+    assertBooleanOrUndefined('progress', progress);
+}
