+++ packages/cspell-lib/src/Settings/DefaultSettings.ts
@@ -121,6 +121,7 @@ export const _defaultSettings: Readonly<CSpellSettingsWithSourceTrace> = {
     ignoreRegExpList: defaultRegExpExcludeList,
     languageSettings: LanguageSettings.getDefaultLanguageSettings(),
     source: { name: 'defaultSettings' },
+    reporters: [],
 };
 
 const getSettings = (function () {
