+++ packages/cspell-lib/src/util/wordSplitter.test.ts
@@ -1,4 +1,4 @@
-import { TextOffset } from './text';
+import { TextOffset } from '@cspell/cspell-types';
 import { SortedBreaks, split, __testing__ } from './wordSplitter';
 
 const generateWordBreaks = __testing__.generateWordBreaks;
