+++ packages/cspell-types/src/index.ts
@@ -1,1 +1,3 @@
 export * from './settings';
+export * from './CSpellReporter';
+export * from './TextOffset';
