+++ packages/cspell/src/index.test.ts
@@ -1,7 +1,8 @@
 import * as index from './index';
 
 // Make sure the types are exported.
-import { CSpellApplicationOptions, RunResult, Emitters, Issue, ProgressFileComplete } from './index';
+import { CSpellApplicationOptions } from './index';
+import { CSpellReporter, Issue, ProgressFileComplete, RunResult } from '@cspell/cspell-types';
 
 describe('Validate index.ts', () => {
     test('index', () => {
@@ -23,7 +24,7 @@ describe('Validate index.ts', () => {
     });
 });
 
-class Logger implements Emitters {
+class Logger implements CSpellReporter {
     log: string[] = [];
     issueCount = 0;
     errorCount = 0;
@@ -31,6 +32,7 @@ class Logger implements Emitters {
     infoCount = 0;
     progressCount = 0;
     issues: Issue[] = [];
+    runResult: RunResult | undefined;
 
     issue = (issue: Issue) => {
         this.issues.push(issue);
@@ -58,5 +60,9 @@ class Logger implements Emitters {
     progress = (p: ProgressFileComplete) => {
         this.progressCount += 1;
         this.log.push(`Progress: ${p.type} ${p.fileNum} ${p.fileCount} ${p.filename}`);
+    };
+
+    result = (r: RunResult) => {
+        this.runResult = r;
     };
 }
