+++ packages/cspell/src/lint.test.ts
@@ -1,6 +1,6 @@
 import { runLint } from './lint';
 import { CSpellApplicationConfiguration } from './CSpellApplicationConfiguration';
-import { Emitters, Issue, ProgressFileComplete } from './emitters';
+import { CSpellReporter, Issue, ProgressFileComplete, RunResult } from '@cspell/cspell-types';
 import * as path from 'path';
 
 const samples = path.resolve(__dirname, '../samples');
@@ -17,7 +17,7 @@ describe('Linter Validation Tests', () => {
     });
 });
 
-class Logger implements Emitters {
+class Logger implements CSpellReporter {
     log: string[] = [];
     issueCount = 0;
     errorCount = 0;
@@ -25,6 +25,7 @@ class Logger implements Emitters {
     infoCount = 0;
     progressCount = 0;
     issues: Issue[] = [];
+    runResult: RunResult | undefined;
 
     issue = (issue: Issue) => {
         this.issues.push(issue);
@@ -52,5 +53,9 @@ class Logger implements Emitters {
     progress = (p: ProgressFileComplete) => {
         this.progressCount += 1;
         this.log.push(`Progress: ${p.type} ${p.fileNum} ${p.fileCount} ${p.filename}`);
+    };
+
+    result = (r: RunResult) => {
+        this.runResult = r;
     };
 }
