+++ test-packages/test-cspell/package.json
@@ -13,7 +13,8 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "cspell": "^5.9.0"
+    "cspell": "^5.9.0",
+    "@cspell/cspell-types": "^5.9.0"
   },
   "main": "index.js",
   "keywords": []
