+++ test-packages/test-cspell/src/index.ts
@@ -1,14 +1,6 @@
 import { assert } from 'console';
-import {
-    checkText,
-    lint,
-    trace,
-    Emitters,
-    Issue,
-    ProgressFileComplete,
-    CSpellApplicationOptions,
-    RunResult,
-} from 'cspell';
+import { checkText, lint, trace, Issue, ProgressFileComplete, CSpellApplicationOptions, RunResult } from 'cspell';
+import { CSpellReporter } from '@cspell/cspell-types';
 import { run } from 'cspell/dist/app';
 
 async function test() {
@@ -35,7 +27,7 @@ async function test() {
     console.log('done');
 }
 
-class ConsoleLogger implements Emitters {
+class ConsoleLogger implements CSpellReporter {
     log: string[] = [];
     issueCount = 0;
     errorCount = 0;
@@ -43,6 +35,7 @@ class ConsoleLogger implements Emitters {
     infoCount = 0;
     progressCount = 0;
     issues: Issue[] = [];
+    runResult: RunResult | undefined;
 
     issue = (issue: Issue) => {
         this.issueCount += 1;
@@ -69,6 +62,10 @@ class ConsoleLogger implements Emitters {
     progress = (p: ProgressFileComplete) => {
         this.progressCount += 1;
         console.error(`Progress: ${p.type} ${p.fileNum} ${p.fileCount} ${p.filename}`);
+    };
+
+    result = (r: RunResult) => {
+        this.runResult = r;
     };
 }
